Refact : Redis 비번 삭제 및 삭제된 파일들 복구

+++ src/main/java/com/example/catchroom_be/domain/product/service/SaleService.java
@@ -9,6 +9,8 @@ import com.example.catchroom_be.domain.product.dto.response.SaleGetDetailInfoRes
 import com.example.catchroom_be.domain.product.dto.response.SaleRegistResponse;
 import com.example.catchroom_be.domain.product.entity.Product;
 import com.example.catchroom_be.domain.product.repository.ProductRepository;
+import com.example.catchroom_be.domain.user.entity.User;
+import com.example.catchroom_be.domain.user.repository.UserEntityRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,29 +20,41 @@ import org.springframework.transaction.annotation.Transactional;
 public class SaleService {
     private final ProductRepository productRepository;
     private final OrderHistoryRepository orderHistoryRepository;
+    private final UserEntityRepository userRepository;
     @Transactional(readOnly = true)
     public SaleGetDetailInfoResponse findProductDetailInfo(Long orderHistroyId) {
         OrderHistory orderHistory = orderHistoryRepository.findById(orderHistroyId).orElseThrow(IllegalArgumentException::new);
         return SaleGetDetailInfoResponse.fromEntity(orderHistory);
     }
     @Transactional
-    public SaleRegistResponse registerProduct(SaleRegistRequest saleRegisterRequest) {
+    public SaleRegistResponse registerProduct(SaleRegistRequest saleRegisterRequest, User user) {
         OrderHistory orderHistory = orderHistoryRepository.findById(saleRegisterRequest.getOrderHistoryId()).orElseThrow(IllegalArgumentException::new);
         orderHistory.updateSaleState(true);
-        Product product = productRepository.save(saleRegisterRequest.toEntity(orderHistory));
+        Product product = productRepository.save(saleRegisterRequest.toEntity(orderHistory,user));
         return SaleRegistResponse.fromEntity(product);
     }
     @Transactional
-    public SaleEditResponse editProduct(Long productId, SaleEditRequest saleEditRequest) {
+    public SaleEditResponse editProduct(Long productId, SaleEditRequest saleEditRequest, User user) {
+        User loginUser = userRepository.findById(user.getId()).orElseThrow(IllegalArgumentException::new);
         Product product = productRepository.findById(productId).orElseThrow(IllegalArgumentException::new);
-        product.updateProduct(saleEditRequest);
+        if (loginUser.getId().equals(product.getSeller().getId())) {
+            product.updateProduct(saleEditRequest);
+        } else {
+            throw new IllegalArgumentException("상품 등록자만 수정할 수 있습니다."); //TODO 예외처리 때 마무리
+        }
         return SaleEditResponse.fromEntity(product);
     }
     @Transactional
-    public void deleteProduct(Long productId) {
+    public void deleteProduct(Long productId, User user) {
+        User loginUser = userRepository.findById(user.getId()).orElseThrow(IllegalArgumentException::new);
         Product product = productRepository.findById(productId).orElseThrow(IllegalArgumentException::new);
-        OrderHistory orderHistory = orderHistoryRepository.findById(product.getOrderHistory().getId()).orElseThrow(IllegalArgumentException::new);
-        orderHistory.updateSaleState(false);
-        productRepository.deleteById(productId);
+        if (loginUser.getId().equals(product.getSeller().getId())) {
+            OrderHistory orderHistory = orderHistoryRepository.findById(product.getOrderHistory().getId()).orElseThrow(IllegalArgumentException::new);
+            orderHistory.updateSaleState(false);
+            productRepository.deleteById(productId);
+        } else {
+            throw new IllegalArgumentException("상품 등록자만 삭제할 수 있습니다."); //TODO 예외처리 때 마무리
+        }
+
     }
 }
